+++ src/gc/gcobject.c
@@ -39,7 +39,7 @@ PObj *NewStrObject(Pgc *gc, Token *name, char *value, bool virt) {
 }
 
 PObj *NewFuncObject(
-    Pgc *gc, Token *name, Token **params, PStmt *body, void *env, size_t count
+    Pgc *gc, Token *name, Token **params, PStmt *body, void *env, u64 count
 ) {
     PObj *o = NewObject(gc, OT_FNC);
     o->v.OFunction.name = name;
@@ -50,7 +50,7 @@ PObj *NewFuncObject(
     return o;
 }
 
-PObj *NewArrayObject(Pgc *gc, Token *op, PValue *items, size_t count) {
+PObj *NewArrayObject(Pgc *gc, Token *op, PValue *items, u64 count) {
     PObj *o = NewObject(gc, OT_ARR);
 
     o->v.OArray.items = items;
