+++ src/include/strescape.h
@@ -5,6 +5,7 @@
 extern "C" {
 #endif
 #include <stddef.h>
+#include "ptypes.h"
 #define ERROR_UNICODE_CP 0xFFFD
 // String Escape Errors
 typedef enum StrEscapeErr {
@@ -38,7 +39,7 @@ typedef enum StrEscapeErr {
 // `outlen` = Length of Output buffer `output`
 // Returns => SESC_OK if nothing goes wrong or other errors from `StrEscapeErr`
 StrEscapeErr ProcessStringEscape(
-    const char *input, size_t inlen, char *output, size_t outlen
+    const char *input, u64 inlen, char *output, u64 outlen
 );
 #ifdef __cplusplus
 }
